+++ components/organisms/place/PlacePreview.tsx
@@ -4,6 +4,7 @@ import { HTMLAttributes, useState } from 'react';
 
 import { MiddlePreviewWrapper } from '@/components/organisms/wrapper/MiddlePreviewWrapper';
 import type { EditorBlock } from '@/shared/types/block';
+import { formatPhoneNumber } from '@/shared/utils/phoneNumber';
 
 import { NaverMapScript } from './NaverMapScript';
 import { Navigation } from './Navigation';
@@ -50,7 +51,9 @@ export const PlacePreview = ({
         </p>
         <p className="font-semibold pb-2.5 text-sm">{placeAddress}</p>
       </section>
-      <p className="font-normal text-text-tertiary pb-2.5">TEL. {placeTel}</p>
+      <p className="font-normal text-text-tertiary pb-2.5">
+        TEL. {formatPhoneNumber(placeTel)}
+      </p>
 
       {blockInfo.props.openMap && isScriptLoaded && (
         <PlaceMap
